Detect slice of object to handle
py::object inherits publicly from py::handle, which means it can slice.

This can easily result in dangling references when using a py::object in a context that expects a py::handle to a new reference, e.g. when returning py::handle from a caster.

To diagnose this, detect when a py::handle is created from a py::object xvalue, preserve the reference count, and decref it when the py::handle is destroyed (only in debug builds); if the refcnt is 1 entering the py::handle destructor it would have been a dangling reference.

As part of this, set m_ptr to nullptr when refcnt is decref'd from 1 to 0.